i'm uploading the thing
a project i made in processing for my introductory mechanics (8.012)
project; it portrays two moons orbiting the Earth--one is the moon in
its current orbit and the other one is in an eccentric elliptical orbit
that brings it very close to the Earth.  I also attempted to display
the tidal bulge and day/night sides of all involved celestial bodies.